Rename `<Drawer>` component to `<Panel>`
To avoid confusion with Drawer related components in the app
