+++ src/components/Panel/PanelLabel.vue
